+++ src/Krystal/Db/Sql/Db.php
@@ -1864,4 +1864,32 @@ final class Db implements DbInterface, RelationableServiceInterface
         $this->queryBuilder->primaryKey($target);
         return $this;
     }
+
+    /**
+     * Appends "CREATE INDEX" expression
+     * 
+     * @param string $name Index name
+     * @param string $table Target table
+     * @param array|string $target A collection of column names, or a single one
+     * @param boolean $unique Whether to include UNIQUE constraint
+     * @return \Krystal\Db\Sql\Db
+     */
+    public function createIndex($table, $name, $target, $unique = false)
+    {
+        $this->queryBuilder->createIndex($table, $name, $target, $unique);
+        return $this;
+    }
+
+    /**
+     * Appends "DROP INDEX" expression
+     * 
+     * @param string $table Target table
+     * @param string $name Index name
+     * @return \Krystal\Db\Sql\Db
+     */
+    public function dropIndex($table, $name)
+    {
+        $this->queryBuilder->dropIndex($table, $name);
+        return $this;
+    }
 }