+++ src/Krystal/Db/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGELOG
  * Dropped `TableDumper` and moved its methods to database service - `dump()` and `fetchAllTables()`
  * Added `getWithWildcart()` in database service. It can be used to build parameters with wildcarts in LIKE queries
  * Removed `TableBuilder`
+ * In database service, added `createIndex()` and `dropIndex()` to create and drop indexes
  * In database service, added `primaryKey()` to create PRIMARY KEYs
  * In database service, added `addConstraint()` and `dropConstraint()`
  * In database service, added `createTable()` method to dynamically create tables